requirements.txt
Django
ipdb
pytest-django
pytest-cov
django-debug-toolbar
psycopg2-binary
Jinja2
uwsgi
django-model-utils
django-filer
django-ordered-model
django-mptt
django-environ
django-tabler
pytest
pytest-cov
pytest-django
pytest-factoryboy
python-dotenv
django-annoying
django-sendfile
whitenoise
